feat(core): introduce data-type-adatpers + date type Introduce a data type adapter to be able to parse FHIR data and expose formaters
